All entities was created and added to database with migrations

+++ API/RestaurantManagement/RestaurantManagement.BLL/Services/Concrete/FileService.cs
@@ -20,7 +20,12 @@ namespace RestaurantManagement.BLL.Services.Concrete
 
         public string AddFile(IFormFile file, string targetDirectory)
         {
-            string uploadsRootFolder = Path.Combine(_hostEnvironment.WebRootPath, targetDirectory);
+            if (string.IsNullOrEmpty(_hostEnvironment.ContentRootPath))
+            {
+                throw new InvalidOperationException("Web root path is not set.");
+            }
+
+            string uploadsRootFolder = Path.Combine(_hostEnvironment.ContentRootPath, targetDirectory);
             if (!Directory.Exists(uploadsRootFolder))
             {
                 Directory.CreateDirectory(uploadsRootFolder);
@@ -34,12 +39,17 @@ namespace RestaurantManagement.BLL.Services.Concrete
                 file.CopyTo(fileStream);
             }
 
-            return $"/{targetDirectory}/{fileName}";
+            return filePath;
         }
 
         public void DeleteFile(string fileName, string targetDirectory)
         {
-            string fileToDeletePath = Path.Combine(_hostEnvironment.WebRootPath, targetDirectory, fileName);
+            if (string.IsNullOrEmpty(targetDirectory))
+            {
+                throw new ArgumentException("Target directory cannot be null or empty.", nameof(targetDirectory));
+            }
+
+            string fileToDeletePath = Path.Combine(_hostEnvironment.ContentRootPath, targetDirectory, fileName);
             if (File.Exists(fileToDeletePath))
             {
                 File.Delete(fileToDeletePath);
